Moved serveral methods from `core.utils` and `input_utils` to `quatrex.device.inputs`
- This was done for the clarity of the package structure.
- Thus, `input_utils` is deleted.
- Tests got moved to `quatrex.device.test_inputs`
- Renamed one test to match the tested function.
