+++ src/quatrex/electron/solver.py
@@ -21,7 +21,8 @@ from quatrex.core.compute_config import ComputeConfig
 from quatrex.core.quatrex_config import QuatrexConfig
 from quatrex.core.statistics import fermi_dirac
 from quatrex.core.subsystem import SubsystemSolver
-from quatrex.core.utils import get_periodic_superblocks, homogenize, load_matrix
+from quatrex.core.utils import get_periodic_superblocks, homogenize
+from quatrex.device.inputs import load_matrix
 
 profiler = Profiler()
